requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
minio==7.1.17
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-dotenv==1.0.0
cryptography==42.0.5
pydantic==2.5.3
aiofiles==23.2.1
